userspace/sdchecks/requirements-native-3.txt
# used by all sdchecks core files
future==0.17.1

# needed for sdchecks.py and config.py
pyyaml==5.1

# checks.d/mcache.py
python-memcached==1.59

# checks.d/php-fpm.py
flup-py3==1.0.3
